Implemented mount_point chown feature
This may be required when mounting a volume for a particular database or
application, that otherwise wouldn't be able to write in the directory when
started, if it's owned by root:root.

Note that this is different from mounting a volume with `-o uid=x,gid=y` as
that option, not present in ext* filesystems, would make all files owned by
the specified uid/gid, which may not necessarily be what we want.

This is how to mandate the owner change for the mount point:

```
mount:
  mount_point:
    path: /mnt
    chown:
      user: postgres
      group: postgres
```

User and group can either be names or numeric IDs, and either of them (but
not both) are optional.